requirements.txt
click==8.1.3
pandas==2.0.2
pandas_profiling==3.2.0
Pillow==9.5.0
Pillow==9.5.0
Requests==2.31.0
streamlit==1.23.1
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.6
streamlit_pandas_profiling==0.1.3
scikit-learn==1.2.2
